feat: extract WebSocket logic into useConversation hook
Create useConversation hook to manage conversation WebSocket connections,
message capture, and transcript processing. This extracts ~400 lines of
complex state management from ConversationInterface, leaving it as a pure
rendering component that consumes the hook.

The hook returns: startConversation, endConversation, messages,
currentSpeaker, isConnecting, isActive

Key changes:
- Created src/hooks/useConversation.ts with full WebSocket/message logic
- Refactored ConversationInterface to use the hook
- Reduced component complexity and improved testability
- All tests pass (9/9 frontend tests, no regressions)
